* fix proxy resolution for artifact resolving: proxy resolution should not be dependant on KobaltContext
